refactor(sync): update 6 consumer imports to use typealias re-export
These synchronizer files import the 4 chunk domain types
(BasicRecord, OcidMapping, CalculatedEquipmentItem, GroupedEquipmentResult)
from their new location via the typealias in
module-synchronizer/.../domain/ChunkDomainReexport.kt.

This is a leftover from Task 2 of issue #1217. The compile was clean
after the typealias was added (compile path uses the new common
ObjectStorage via the typealias). These 6 files were part of the
typealias import migration.

Issue #1217
