+++ module-synchronizer/src/main/kotlin/maple/synchronizer/redis/OcidMappingRedisWriter.kt
@@ -1,6 +1,6 @@
 package maple.synchronizer.redis
 
-import maple.synchronizer.storage.OcidMapping
+import maple.synchronizer.domain.OcidMapping
 import org.slf4j.LoggerFactory
 import org.springframework.data.redis.core.StringRedisTemplate
 import org.springframework.stereotype.Component
